+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added 
 - `MLBMatchupComparison` for fox sports json response for probable starting pitchers
+- `MLBProbableStartingPitcher` fox sports data model for probable starting pitcher
+- Scraping fox sports MLB probable starting pitcher using `MLBProbableStartingPitcherScraper`
 
 ## [0.8.2] - 2025-05-17
 ### Fixed